functions/redis-function/requirements.txt
redis==4.4.0
Flask==2.2.2
google-cloud-logging==2.7.0
